requirements.txt
firebase-admin
openai_api
ffmpeg
flask
flask_cors
ipdb